+++ faz/resource_fmupdate_fdssetting.go
@@ -190,6 +190,11 @@ func resourceFmupdateFdsSetting() *schema.Resource {
 					},
 				},
 			},
+			"system_support_fai": &schema.Schema{
+				Type:     schema.TypeSet,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Optional: true,
+			},
 			"system_support_faz": &schema.Schema{
 				Type:     schema.TypeSet,
 				Elem:     &schema.Schema{Type: schema.TypeString},
@@ -624,6 +629,10 @@ func flattenFmupdateFdsSettingServerOverrideStatusFfa(v interface{}, d *schema.R
 	return v
 }
 
+func flattenFmupdateFdsSettingSystemSupportFaiFfa(v interface{}, d *schema.ResourceData, pre string) interface{} {
+	return flattenStringList(v)
+}
+
 func flattenFmupdateFdsSettingSystemSupportFazFfa(v interface{}, d *schema.ResourceData, pre string) interface{} {
 	return flattenStringList(v)
 }
@@ -903,6 +912,16 @@ func refreshObjectFmupdateFdsSetting(d *schema.ResourceData, o map[string]interf
 					return fmt.Errorf("Error reading server_override: %v", err)
 				}
 			}
+		}
+	}
+
+	if err = d.Set("system_support_fai", flattenFmupdateFdsSettingSystemSupportFaiFfa(o["system-support-fai"], d, "system_support_fai")); err != nil {
+		if vv, ok := fortiAPIPatch(o["system-support-fai"], "FmupdateFdsSetting-SystemSupportFai"); ok {
+			if err = d.Set("system_support_fai", vv); err != nil {
+				return fmt.Errorf("Error reading system_support_fai: %v", err)
+			}
+		} else {
+			return fmt.Errorf("Error reading system_support_fai: %v", err)
 		}
 	}
 
@@ -1313,6 +1332,10 @@ func expandFmupdateFdsSettingServerOverrideStatusFfa(d *schema.ResourceData, v i
 	return v, nil
 }
 
+func expandFmupdateFdsSettingSystemSupportFaiFfa(d *schema.ResourceData, v interface{}, pre string) (interface{}, error) {
+	return expandStringList(v.(*schema.Set).List()), nil
+}
+
 func expandFmupdateFdsSettingSystemSupportFazFfa(d *schema.ResourceData, v interface{}, pre string) (interface{}, error) {
 	return expandStringList(v.(*schema.Set).List()), nil
 }
@@ -1535,6 +1558,15 @@ func getObjectFmupdateFdsSetting(d *schema.ResourceData) (*map[string]interface{
 			return &obj, err
 		} else if t != nil {
 			obj["server-override"] = t
+		}
+	}
+
+	if v, ok := d.GetOk("system_support_fai"); ok || d.HasChange("system_support_fai") {
+		t, err := expandFmupdateFdsSettingSystemSupportFaiFfa(d, v, "system_support_fai")
+		if err != nil {
+			return &obj, err
+		} else if t != nil {
+			obj["system-support-fai"] = t
 		}
 	}